+++ controllers/servicioCtrl.php
@@ -14,7 +14,10 @@
 			switch ($_GET['act']) {
 				case 'create':
 					//Crear 
-					$this->create();
+					if(BaseCtrl::isAdmin())
+						$this->create();
+					else
+						require_once 'views/permisosError.html';
 					break;
 				case 'lists':
 					//Listar 
@@ -22,11 +25,17 @@
 					break;
 				case 'delete':
 					//Baja 
-					$this->delete();
+					if(BaseCtrl::isAdmin())
+						$this->delete();
+					else
+						require_once 'views/permisosError.html';
 					break;
 				case 'update':
 					//Baja
-					$this->update();
+					if(BaseCtrl::isAdmin())
+						$this->update();
+					else
+						require_once 'views/permisosError.html';
 					break;
 				default:
 					# code...
